small bug fixes and corrections

+++ Lab1/LibraryLab1/ConsoleApp/Program.cs
@@ -5,7 +5,7 @@ public class Program
     {
         List<Animal> animals = new List<Animal> { new Mammal("Lion", "11Df12", false), new Bird("Parrot", "31Bf32", true), new Reptile("Snake", "24Vf23") };
         List<ZooStaff> employees = new List<ZooStaff> { new ZooStaff("John", "Briks", "Manager"), new ZooStaff("Alice","Watson", "Veterinarian") };
-        double feedStocks = 123;
+        int feedStocks = 123;
 
         Inventory inventory = new Inventory(animals, employees, feedStocks);
         inventory.ShowAnimalInventory();
